Address L2 review: type console timestamp_ms as number, not bigint
ConsoleRecord.timestamp_ms was a Rust u64, which weld statically maps to
TS `bigint` (u64 can exceed 2^53). But serde_v8 serializes a u64 as a
v8::Number whenever it is <= 2^53-1 (only larger values become a BigInt),
and a millisecond timestamp stays well under that bound until ~year
287396. So the value actually arrives in JS as a `number`, and the
`bigint` SDK/TS type misrepresented it.

Fix at the root so the Rust type, generated SDK, hand-written TS, and the
runtime value all agree on `number`: change the field to f64 (which weld
maps to `number`, serde_v8 emits unconditionally as v8::Number, and which
matches JavaScript's native time representation -- Date.now() is a
number). now_millis() and build_record() now use f64, and
parse_renderer_message reads the timestamp via Value::as_f64. The
regenerated sdk/runtime.console.ts and ts/init.ts both type
timestamp_ms as `number`.

cargo test -p ext_console (14) / clippy / fmt all clean.

+++ crates/ext_console/src/console_log.rs
@@ -45,7 +45,12 @@ pub struct ConsoleRecord {
     /// The arguments passed to the console call, preserved as JSON values.
     pub args: Vec<Value>,
     /// Wall-clock capture time in milliseconds since the UNIX epoch.
-    pub timestamp_ms: u64,
+    ///
+    /// Stored as `f64` to match JavaScript's native time representation
+    /// (`Date.now()` is a `number`): millisecond timestamps are exact in `f64`
+    /// well beyond any realistic date, and serde_v8 serializes safe-range
+    /// values as a JS `number` rather than a `bigint`.
+    pub timestamp_ms: f64,
     /// Whether the record came from the Deno side or the renderer.
     pub source: ConsoleSource,
 }
@@ -117,7 +122,7 @@ mod tests {
         ConsoleRecord {
             level: level.to_string(),
             args: vec![json!(msg)],
-            timestamp_ms: 0,
+            timestamp_ms: 0.0,
             source: ConsoleSource::Deno,
         }
     }
